+++ src/pyradigms/__init__.py
@@ -1,16 +1,29 @@
+"""Docstring describing pyradigms"""
 import pandas as pd
 import numpy as np
 import re
 from io import StringIO
 from clldutils.loglib import get_colorlog
 import sys
+from pathlib import Path
 from attrs import define, Factory
 from typing import List, Dict
+import logging
+import colorlog
 
-"""Docstring describing pyradigms"""
+
+handler = colorlog.StreamHandler(None)
+handler.setFormatter(
+    colorlog.ColoredFormatter("%(log_color)s%(levelname)-7s%(reset)s %(message)s")
+)
+log = logging.getLogger(__name__)
+log.setLevel(logging.DEBUG)
+log.propagate = True
+log.addHandler(handler)
+
 
 person_values = ["1", "2", "3", "1+3", "1+2"]
-content_string = "Form"
+print_column = "Form"
 # separators = ["."]
 x = []
 y = []
@@ -20,6 +33,44 @@ y_sort = []
 sort_orders = {}
 filters = {}
 ignore = []
+
+
+def format_person_values(string, sep):
+    for val in person_values:
+        if val in string:
+            string = string.replace(val + sep, val)
+    return string
+
+
+# cast parameters to list
+def listify(var):
+    if not isinstance(var, list) and not isinstance(var, tuple):
+        return [var]
+    return var
+
+
+def get_parameter_values(string, parameters, separators):
+    parameter_list = re.split("|".join(map(re.escape, separators)), string)
+    parameter_list = [x for x in parameter_list if x]  # Remove leftovers of separation
+    new_parameter_list = []
+    for param in parameter_list:
+        p_hit = False
+        for p_v in sorted(person_values, key=len, reverse=True):
+            if param.startswith(p_v):
+                new_parameter_list += list(filter(None, list(param.partition(p_v))))
+                p_hit = True
+                break
+        if not p_hit:
+            new_parameter_list.append(param)
+    parameter_list = new_parameter_list
+    if len(parameter_list) < len(parameters):
+        for i in range(0, len(parameters) - len(parameter_list)):
+            del i
+            parameter_list.append(None)
+    elif len(parameter_list) > len(parameters):
+        log.error(f"More values than specified: {parameters} {parameter_list}")
+        sys.exit(1)
+    return parameter_list
 
 
 @define
@@ -38,26 +89,26 @@ class Pyradigm:
     sort_orders: Dict = Factory(dict)
     filters: Dict = Factory(dict)
     ignore: List[str] = Factory(list)
-    separators: List[str] = Factory(list)
+    with_multi_index = False
+    separators = ["."]
     log_level: str = None
     logger = get_colorlog(__name__, sys.stdout, level="DEBUG")
-    joiner = "/"
-    content_string: str = "Form"
-
-    def __attrs_post_init__(self):
-        if self.log_level is not None:
-            self.logger.setLevel(self.log_level)
-        if self.separators is None or self.separators == []:
-            self.separators = ["."]
+    value_joiner = "."
+    category_joiner = " / "
+    print_column = "Form"
+    output_folder = None
 
     @property
     def parameters(self):
         parlist = list(self.entries.columns)
-        parlist.remove(self.content_string)
+        parlist.remove(self.print_column)
         return parlist
 
+    def _short_repr(self, df):
+        return df.head().to_string() + f"\n({len(df)} entries)"
+
     @classmethod
-    def from_dataframe(cls, df: pd.DataFrame, format="wide", **kwargs):
+    def from_dataframe(cls, df: pd.DataFrame, data_format="wide", **kwargs):
         """Create a new Pyradigm from a pandas dataframe.
 
         :param df: a pandas dataframe containing the data in wide/unstacked format
@@ -65,18 +116,21 @@ class Pyradigm:
         :return: a Pyradigm object
         """
 
-        if format == "wide":
+        if data_format == "wide":
             return cls(entries=df, **kwargs)
-        elif format == "long":
-            print(df)
-            return df
-        elif format == "paradigm":
+        if data_format == "long":
+            out = df.pivot(index="ID", columns="Parameter", values="Value")
+            out.columns.name = None
+            return cls(entries=out, **kwargs)
+        if data_format == "paradigm":
             return cls(
                 entries=cls.decompose_paradigm(cls, paradigm=df, **kwargs), **kwargs
             )
+        log.error(f"Invalid format: {data_format}")
+        sys.exit(1)
 
     @classmethod
-    def from_csv(cls, path, format="wide", **kwargs):
+    def from_csv(cls, path, data_format="wide", **kwargs):
         """Create a new Pyradigm from a CSV file.
 
         :param path: a path to a CSV file containing the data in wide/unstacked format
@@ -84,42 +138,18 @@ class Pyradigm:
         :return: a Pyradigm object
         """
         df = pd.read_csv(path, keep_default_na=False, dtype=str)
-        if format == "wide":
+        if data_format == "wide":
             return cls(entries=df, **kwargs)
-        elif format == "long":
-            print(df)
-            print(df.pivot(index="Form", columns="Parameter")["Value"])
-        elif format == "paradigm":
+        if data_format == "long":
+            # print(df.pivot(index="Form", columns="Parameter")["Value"])
+            return df
+        if data_format == "paradigm":
             df.set_index(df.columns[0], inplace=True)
             return cls(
                 entries=cls.decompose_paradigm(cls, paradigm=df, **kwargs), **kwargs
             )
-
-    def decompose_paradigm(
-        self, paradigm, x, y, separators=["."], z_value=None, **kwargs
-    ):
-        if not z_value:
-            z_value = paradigm.index.name
-
-        entries = pd.DataFrame(columns=z + x + y + [content_string])
-        for i, (x_string, row) in enumerate(paradigm.iteritems()):
-            for y_string, form in row.iteritems():
-                x_values = self.get_parameter_values(x_string, x, separators)
-                y_values = self.get_parameter_values(y_string, y, separators)
-                param_list = x + y  # List of parameter names
-                value_list = x_values + y_values  # List of parameter values
-                if z:
-                    param_list += z
-                    value_list += [z_value]
-                param_list += [content_string]
-                value_list += [form]
-                out_dict = dict(zip(param_list, value_list))
-                entries = pd.concat([entries, pd.DataFrame(out_dict, index=[i])])
-
-        entries.dropna(subset=[content_string], inplace=True)  # …drop rows with no form
-        entries.reset_index(drop=True, inplace=True)  # reset index
-        entries.fillna("", inplace=True)
-        return entries
+        log.error(f"Invalid format: {data_format}")
+        sys.exit(1)
 
     @classmethod
     def from_text(cls, text, x_sep=",", y_sep="\n"):
@@ -134,62 +164,79 @@ class Pyradigm:
         df = pd.read_csv(StringIO(text), sep=x_sep, lineterminator=y_sep)
         return cls(df)
 
-    def to_markdown(self, format="paradigm", **kwargs):
-        if format == "long":
+    def decompose_paradigm(self, paradigm, z_value=None, **kwargs):
+        """Important: the y axis labels must be the index, not the first column"""
+        x = listify(kwargs.get("x", self.x))
+        y = listify(kwargs.get("y", self.y))
+        z = listify(kwargs.get("z", self.z))
+        separators = kwargs.get("separators", self.separators)
+        print_column = kwargs.get("print_column", self.print_column)
+
+        log.error(print_column)
+        if not z_value:
+            z_value = paradigm.index.name
+
+        # gather all parameter names from the defined axes + the name of what's in the cells
+        entries = pd.DataFrame(columns=z + x + y + [print_column])
+
+        for i, (x_string, col) in enumerate(paradigm.iteritems()):
+            for y_string, form in col.iteritems():
+                x_values = get_parameter_values(x_string, x, separators)
+                y_values = get_parameter_values(y_string, y, separators)
+                param_list = x + y  # List of parameter names
+                value_list = x_values + y_values  # List of parameter values
+                if z:
+                    param_list += z
+                    value_list += [z_value]
+                param_list += [print_column]
+                value_list += [form]
+                out_dict = dict(zip(param_list, value_list))
+                entries = pd.concat([entries, pd.DataFrame(out_dict, index=[i])])
+
+        entries.dropna(subset=[print_column], inplace=True)  # …drop rows with no form
+        entries.reset_index(drop=True, inplace=True)  # reset index
+        entries.fillna("", inplace=True)
+        return entries
+
+    def to_markdown(self, data_format="paradigm", **kwargs):
+        if data_format == "long":
             out = self.to_long()
-        elif format == "paradigm":
+        elif data_format == "paradigm":
             out = self.compose_paradigm(**kwargs)
-        elif format == "wide":
+        elif data_format == "wide":
             out = self.entries
+        else:
+            log.error(f"Unknown format {data_format}.")
+            sys.exit(1)
         return out.to_markdown()
 
     def to_long(self):
+        if "ID" not in self.entries:
+            self.entries["ID"] = self.entries.apply(
+                lambda x: f"{x.name}-{x[print_column]}", axis=1
+            )
         return pd.melt(
             self.entries,
-            id_vars=self.content_string,
-            value_vars=self.parameters,
+            id_vars="ID",
+            value_vars=self.parameters + [self.print_column],
             var_name="Parameter",
             value_name="Value",
-        ).drop_duplicates()
+        )
 
-    def compose_paradigm(
-        self,
-        input_df=None,
-        csv_output=None,
-        multi_index=False,
-        x=None,
-        y=None,
-        z=None,
-        x_sort=None,
-        y_sort=None,
-        filters=None,
-        ignore=None,
-        separators=None,
-        joiner=None,
-        content_string=None,
-    ):
-        if input_df is None:
-            input_df = self.entries
-        if x is None:
-            x = self.x
-        if y is None:
-            y = self.y
-        if z is None:
-            z = self.z
-        if x_sort is None:
-            x_sort = self.x_sort
-        if y_sort is None:
-            y_sort = self.y_sort
-        if filters is None:
-            filters = self.filters
-        if ignore is None:
-            ignore = self.ignore
-        if separators is None:
-            separators = self.separators
-        if joiner is None:
-            joiner = self.joiner
-        if content_string is None:
-            content_string = self.content_string
+    def compose_paradigm(self, csv_output=None, **kwargs):
+        input_df = kwargs.get("input_df", self.entries)
+        with_multi_index = kwargs.get("with_multi_index", self.with_multi_index)
+        x = kwargs.get("x", self.x)
+        y = kwargs.get("y", self.y)
+        z = kwargs.get("z", self.z)
+        filters = kwargs.get("filters", self.filters)
+        ignore = kwargs.get("ignore", self.ignore)
+        separators = kwargs.get("separators", self.separators)
+        value_joiner = kwargs.get("value_joiner", self.value_joiner)
+        category_joiner = kwargs.get("category_joiner", self.category_joiner)
+        print_column = kwargs.get("print_column", self.print_column)
+        sort_orders = kwargs.get("sort_orders", self.sort_orders)
+        output_folder = kwargs.get("output_folder", self.output_folder)
 
         df = input_df.copy()
         df.replace(np.nan, "", inplace=True)
@@ -197,12 +244,12 @@ class Pyradigm:
             f"Composing a new paradigm from entries:\n{self._short_repr(df)}"
         )
 
-        # cast parameters to list
-        def listify(var):
-            if type(var) != list:
-                return [var]
-            else:
-                return var
+        # get a sensible default sort order for a given parameter (order in the input)
+        def get_sort_order(parameter):
+            val_list = list(df[parameter])
+            val_list = [i for n, i in enumerate(val_list) if i not in val_list[:n]]
+            log.debug(f"New sort order for {parameter}: {val_list}")
+            return val_list
 
         x = listify(x)
         y = listify(y)
@@ -214,34 +261,43 @@ class Pyradigm:
             remainders = set(v) - set(df.columns)
             if len(remainders) > 0:
                 rstring = ", ".join(remainders)
-                self.logger.error(
-                    f"{k} axis contains inexistent parameter(s): {rstring}"
-                )
-                return None
+                log.error(f"{k} axis contains inexistent parameter(s): {rstring}")
+                sys.exit(1)
 
-        # make sure that "Form" or whatever other content string is present
-        if content_string not in df.columns:
-            self.logger.error(
-                f"String {content_string} is not found in dataframe columns"
+        log.info(print_column in df.columns)
+        # make sure that "Form" or whatever other column to print is present
+        if print_column not in df.columns:
+            log.error(f"'{print_column}' not found in dataframe columns")
+            sys.exit(1)
+
+        # inform user if there are columns they did not give directions for
+        leftover_columns = set(df.columns) - set(x + y + z + [print_column])
+        if len(leftover_columns) > 0:
+            log.info(
+                "You did not specify what should happen to the following columns/fields/parameters: %s",
+                "\t".join(leftover_columns),
             )
-            return None
 
+        # only for debugging purposes
         filter_string = "\n".join(
             [f"\t{k}: {', '.join(v)}" for k, v in filters.items()]
         )
-        self.logger.debug(f"Filtering parameters:\n{filter_string}\n")
-        for col, values in filters.items():
-            if type(values) != list:
-                values = [values]
-            df = df[df[col].isin(values)]
-        self.logger.debug("New entries:\n" + self._short_repr(df))
+        log.debug(f"Filtering parameters:\n{filter_string}\n")
 
+        # filter rows by fitler arg
+        for col, values in filters.items():
+            values = listify(values)
+            df = df[df[col].isin(values)]
+        log.debug("Filtered entries:\n%s", self._short_repr(df))
+
+        # drop irrelevant columns
         if len(ignore) > 0:
-            self.logger.debug(f"""Ignoring parameters:\n{", ".join(ignore)}""")
+            log.debug(f"""Ignoring parameters:\n{", ".join(ignore)}""")
             for col in ignore:
                 df.drop(col, axis=1, inplace=True)
-                self.logger.debug("New entries:\n" + self._short_repr(df))
+                log.debug("New entries:\n%s", self._short_repr(df))
 
+        # join columns for the y axis
         def concat_values(row, values):
             out = ""
             for value in values:
@@ -252,19 +308,9 @@ class Pyradigm:
                 separators[0] + separators[0], separators[0]
             )
 
-        new_y_id = separators[0].join(y)
-        self.logger.debug(f"New y id: [{new_y_id}]")
-        if len(y) > 1 and not multi_index:
-            df[new_y_id] = df.apply(concat_values, values=y, axis=1)
-
-        new_x_id = separators[0].join(x)
-        self.logger.debug(f"New x id: [{new_x_id}]")
-        if len(x) > 1 and not multi_index:
-            df[new_x_id] = df.apply(concat_values, values=x, axis=1)
-
         new_z_id = separators[0].join(z)
         if len(z) > 1:
-            self.logger.debug(f"New z id: [{new_z_id}]")
+            # self.logger.debug(f"New z id: [{new_z_id}]")
             df[new_z_id] = df.apply(concat_values, values=z, axis=1)
 
         if len(z) > 0:
@@ -275,38 +321,19 @@ class Pyradigm:
         constructed_paradigms = {}
         pd.set_option("display.max_rows", None, "display.max_columns", None)
         for z_key, df in z_dict.items():
-            self.logger.debug(
-                f"Creating pivot table for [{z_key}], x: [{new_x_id}], y: [{new_y_id}]"
+            log.debug(
+                f"Creating pivot table for x={x}, y={y}, cell values: {print_column}"
             )
-            if multi_index:
-                out = pd.pivot_table(
-                    df,
-                    values=content_string,
-                    index=y,
-                    columns=x,
-                    aggfunc=lambda x: joiner.join(x),
-                )
-            else:
-                out = pd.pivot_table(
-                    df,
-                    values=content_string,
-                    index=new_y_id,
-                    columns=new_x_id,
-                    aggfunc=lambda x: joiner.join(x),
-                )
-            if not multi_index:
-                out.rename_axis(
-                    None, axis="columns", inplace=True
-                )  # remove column labels
-            if len(z_dict) == 1:
-                out.index.name = new_y_id
-            else:
-                out.index.name = z_key
+            out = pd.pivot_table(
+                df,
+                values=print_column,
+                index=y,
+                columns=x,
+                aggfunc=category_joiner.join,
+            )
 
-            if multi_index:
-                idx_name = out.index.names
-            else:
-                idx_name = out.index.name
+            # drop empty rows
+            idx_name = out.index.names
             out.reset_index(inplace=True)  # use index as column
             out.replace(
                 "", np.nan, inplace=True
@@ -316,91 +343,128 @@ class Pyradigm:
                 "", inplace=True
             )  # then add back the empty strings for exporting
             out.set_index(idx_name, drop=True, inplace=True)  # then add back the index
-            out.index.name = ""
-            if not multi_index:
-                out = out.reindex(
-                    [value for value in y_sort if value in out.index]
-                    + list(set(list(out.index)) - set(y_sort))
-                )  # sort index by specified order, put leftovers at the end
-                comp_x_sort = x_sort + list(
-                    set(list(out.columns)) - set(x_sort)
-                )  # sort columns, too
-                comp_x_sort = dict(zip(comp_x_sort, range(len(comp_x_sort))))
-                out = out[sorted(out.columns, key=lambda x: comp_x_sort[x])]
-            else:
-                new_indices = []
-                for lvl_c in range(0, out.index.nlevels):
-                    values = out.index.get_level_values(y[lvl_c])
-                    if y[lvl_c] in sort_orders:
-                        new_indices.append(
-                            pd.CategoricalIndex(
-                                values, categories=sort_orders[y[lvl_c]], ordered=True
-                            )
-                        )
-                    else:
-                        new_indices.append(pd.CategoricalIndex(values))
-                out.set_index(new_indices, inplace=True)
-                out.sort_index(inplace=True)
+            # out.index.name = ""
 
-                new_columns = []
-                for lvl_c in range(0, out.columns.nlevels):
-                    values = out.columns.get_level_values(x[lvl_c])  # != ""
-                    if x[lvl_c] in sort_orders:
-                        new_columns.append(
-                            pd.CategoricalIndex(
-                                values, categories=sort_orders[x[lvl_c]], ordered=True
-                            )
-                        )
-                    else:
-                        new_columns.append(pd.CategoricalIndex(values))
-                out.columns = new_columns
-                out.sort_index(inplace=True, axis=1)
+            log.debug(out.index.names)
+            log.debug(out.columns.names)
+            # if not 1==1:
+            #     out = out.reindex(
+            #         [value for value in y_sort if value in out.index]
+            #         + list(set(list(out.index)) - set(y_sort))
+            #     )  # sort index by specified order, put leftovers at the end
+            #     comp_x_sort = x_sort + list(
+            #         set(list(out.columns)) - set(x_sort)
+            #     )  # sort columns, too
+            #     comp_x_sort = dict(zip(comp_x_sort, range(len(comp_x_sort))))
+            #     out = out[sorted(out.columns, key=lambda x: comp_x_sort[x])]
+            # else:
+
+            # for those parameters lacking a specified sort order, establish a default
+            for parameter in out.index.names + out.columns.names:
+                print(parameter)
+                df_sort = get_sort_order(parameter)
+                if parameter not in sort_orders:
+                    sort_orders[parameter] = df_sort
+                elif set(sort_orders[parameter]) != set(df_sort):
+                    log.error(
+                        f"Specified order {sort_orders[parameter]} for parameter '{parameter}' does not cover all values: {df_sort}"
+                    )
+                    sys.exit(1)
+
+            # sort x and y axis
+            new_indices = []
+            for idx in y:
+                values = out.index.get_level_values(idx)
+                log.info(values)
+                new_indices.append(
+                    pd.CategoricalIndex(
+                        values, categories=sort_orders[idx], ordered=True
+                    )
+                )
+            out.set_index(new_indices, inplace=True)
+            out.sort_index(level=[c for c in sort_orders if c in y], inplace=True)
+
+            new_columns = []
+            for col in x:
+                values = out.columns.get_level_values(col)
+                log.info(values)
+                new_columns.append(
+                    pd.CategoricalIndex(
+                        values, categories=sort_orders[col], ordered=True
+                    )
+                )
+            out.columns = new_columns
+            out.sort_index(
+                level=[c for c in sort_orders if c in x], inplace=True, axis=1
+            )
+
+            log.warning(out.columns.values)
+            if not with_multi_index:
+                log.debug("Flattening multiindices")
+
+                new_colindex_name = category_joiner.join(x)
+                out.columns = [
+                    format_person_values(value_joiner.join(col).strip(), value_joiner)
+                    for col in out.columns.values
+                ]
+                out.columns.name = new_colindex_name
+
+                new_index_name = category_joiner.join(y)
+                out.index = [
+                    format_person_values(
+                        value_joiner.join(listify(col)).strip(), value_joiner
+                    )
+                    for col in out.index.values
+                ]
+                out.index.name = new_index_name
+            log.warning(out)
 
             constructed_paradigms[z_key] = out
-        if csv_output:
-            output = []
-            for df in constructed_paradigms.values():
-                s = StringIO()
-                df.to_csv(s, index=True)
-                output.append(s.getvalue())
-            with open(csv_output, "w") as file:
-                file.write("\n".join(output))
-        if len(constructed_paradigms) == 1:
-            return constructed_paradigms["z"]
-        else:
-            return constructed_paradigms
 
-    def get_parameter_values(self, string, parameters, separators):
-        parameter_list = re.split("|".join(map(re.escape, separators)), string)
-        parameter_list = [
-            x for x in parameter_list if x
-        ]  # Remove leftovers of separation
-        new_parameter_list = []
-        for param in parameter_list:
-            p_hit = False
-            for p_v in sorted(person_values, key=len, reverse=True):
-                if param.startswith(p_v):
-                    new_parameter_list += list(filter(None, list(param.partition(p_v))))
-                    p_hit = True
-                    break
-            if not p_hit:
-                new_parameter_list.append(param)
-        parameter_list = new_parameter_list
-        if len(parameter_list) < len(parameters):
-            for i in range(0, len(parameters) - len(parameter_list)):
-                parameter_list.append(None)
-        elif len(parameter_list) > len(parameters):
-            print("ERROR")
-            print(parameter_list)
-            print(parameters)
-        return parameter_list
+        if csv_output:
+            log.debug(f"Writing to {csv_output}")
+            output = []
+            for z_key, df in constructed_paradigms.items():
+                s = StringIO()
+                if not with_multi_index:
+                    if z_key == "z":
+                        idx_label = ""
+                    else:
+                        idx_label = z_key
+                    df.to_csv(s, index=True, index_label=idx_label)
+                else:
+                    df.to_csv(s, index=True)
+                output.append(s.getvalue())
+            with open(csv_output, "w", encoding="utf-8") as file:
+                file.write("\n".join(output))
+
+        if output_folder:
+            log.debug(f"Saving CSV files to {output_folder}")
+            for z_key, df in constructed_paradigms.items():
+                if z_key == "z":
+                    filename = "generated_paradigm.csv"
+                    idx_label = ""
+                else:
+                    filename = z_key + ".csv"
+                    idx_label = z_key
+                if with_multi_index:
+                    df.to_csv(output_folder / filename, index=True)
+                else:
+                    df.to_csv(
+                        output_folder / filename, index=True, index_label=idx_label
+                    )
+
+        if len(constructed_paradigms) == 1:
+            return list(constructed_paradigms.values())[0]
+        return constructed_paradigms
 
     def decompose_from_csv(self, csvfile):
-        dfull = open(csvfile).read()
+        with open(csvfile, "r", encoding="utf-8") as f:
+            dfull = f.read()
         return self.decompose_from_text(dfull)
 
     def decompose_from_text(self, string):
-        full = pd.DataFrame(columns=x)
+        full = pd.DataFrame()
         strings = string.split("\n\n")
         for df_string in strings:
             paradigm = pd.read_csv(StringIO(df_string), dtype=str, index_col=0)
@@ -412,14 +476,3 @@ class Pyradigm:
             full = full.append(self.decompose_paradigm(paradigm, z_value=z_value))
         full.reset_index(drop=True, inplace=True)
         return full
-
-    def compose_from_text(self, df_string, csv_output=None):
-        entries = pd.read_csv(StringIO(df_string), dtype=str, keep_default_na=False)
-        return self.compose_paradigm(entries, csv_output=csv_output)
-
-    def compose_from_csv(self, csvfile, csv_output=None):
-        entries = pd.read_csv(csvfile, dtype=str, keep_default_na=False)
-        return self.compose_paradigm(entries, csv_output=csv_output)
-
-    def _short_repr(self, df):
-        return df.head().to_string() + f"\n({len(df)} entries)"